scraper and controller files rearranged, data response populating correctly for cars.com, will work on populating the database and making scrapers for other sites. Then just need to figure out how we want to display on frontend.
