+++ server/controllers/pgController.js
@@ -1,7 +1,8 @@
 const db = require('../models/pgModels.js');
 const puppeteer = require('puppeteer');
-const axios = require('axios');
-const cheerio = require('cheerio');
+// carsDotComScraper = require('../scrapers/carsDotComScraper')
+const carsDotComScraper = require('../scrapers/carsDotComScraper.js')
+
 
 const pgController = {};
 
@@ -44,83 +45,23 @@ const pgController = {};
 //  scrape();
 
 // }
-
-
-pgController.scrapeCarInfo = (req, res, next) => {
-    // if data exists already in database by the unique listing URL, pull the data from database
-  const {make, model, minYear, zip} = req.body;
-  //cheerioScrapeCarsCom one is working perfeclty
-  cheerioScrapeCarsCom = (url) => {
-    axios(url)
-    .then(response => {
-      
-      const htmlData = response.data;
-      const $ = cheerio.load(htmlData);
-      
-        const cars = [];
-        const scrape = $('.vehicle-cards').find('.vehicle-card').map((i, el) => {
-        
-          const vehicleObj = {};
-          const priceElement = $(el).find('.primary-price');
-          const mileageElement = $(el).find('.mileage');
-          const image = $(el).find('.image-wrap').find('img').attr('data-src');
-          const titleElement = $(el).find('.title'); // title '2015 Honda Civic LX'
-          const url = `cars.com${$(el).find('a').attr('href')}`;
-          vehicleObj.num = i;
-          vehicleObj.price = Number(priceElement.text().replace(/\D/g, ''));
-          vehicleObj.image = image;
-          vehicleObj.mileage = Number(mileageElement.text().replace(/\D/g, ''));
-          [vehicleObj.year, vehicleObj.make, vehicleObj.model] = titleElement.text().split(' '); // [2015, Honda, Civic, LX]
-          vehicleObj.year = Number(vehicleObj.year);
-          vehicleObj.url = url;
-          cars.push(vehicleObj);
-    })
-
-      const date = new Date();
-      const dayDate = date.getDate();
-      const monthDate = date.getMonth() + 1;
-      const yearDate = date.getFullYear();
-      res.locals.carData = {
-        cars,
-        date: `${monthDate}/${dayDate}/${yearDate}`
-      }
-      return next();
-    })
-    .catch(err => console.log(err));
-  }
-  
-  // cheerioScrapeCarsCom is working perfectly
-  cheerioScrapeCarsCom(`https://www.cars.com/shopping/results/?dealer_id=&keyword=&list_price_max=&list_price_min=&makes[]=${make.toLowerCase()}&maximum_distance=200&mileage_max=&models[]=${make.toLowerCase()+'-'+model.toLowerCase()}&page_size=10&sort=list_price&stock_type=all&year_max=&year_min=${minYear}&zip=${zip}`);
-
-  // //I haven't found a website that allows me get data and has a fancy url to work with
-  // cheerioScrapeEdmunds = (url) => {
-  //   axios(url)
-  //     .then(response => {
-  //       const htmlData = response.data;
-  //       const $ = cheerio.load(htmlData);
-  //       const price = $('span[itemprop="price"]').html();
-  //       console.log(price);
-  //       res.status(200).send('done')
-  //       // .slice(1).replace(/\D/g, '') const mileage = $('.vehicle-card .vehicle-card-main .vehicle-details .mileage').html().replace(/\D/g, '');
-  //       // const [year, make, model] = $('.title').html().split(' ');
-  //       // console.log(price, year, make, model, mileage);
-  //       // const date = new Date();
-  //       // const dayDate = date.getDate();
-  //       // const monthDate = date.getMonth() + 1;
-  //       // const yearDate = date.getFullYear();
-  //       // res.locals.carData = {
-  //       //   price, year, make, model, mileage, url,
-  //       //   date: `${monthDate}/${dayDate}/${yearDate}`
-  //       // }
-  //       // return next();
-  //     })
-  //     .catch(err => console.log(err))
-  // }
-  
-  //I haven't found a website that allows me get data and has a fancy url to work with
-  // cheerioScrapeEdmunds(`https://www.carsdirect.com/used_cars/listings/honda/civic?zipcode=33157&dealerId=&distance=&yearFrom=&yearTo=&priceFrom=&priceTo=&qString=Honda%603%6023%600%600%60false%7CCivic%604%60253%600%600%60false%7C&keywords=&makeName=honda&modelName=civic&sortColumn=Price&sortDirection=ASC&searchGroupId=&lnk=&vehicleDetailLead=false&recentSearchId=12762496&pageNum=1`);
+pgController.getCarsComData = async (req, res, next) => {
+ const { make, model, minYear, zip } = req.body;
+ res.locals.carsComData = await carsDotComScraper(make, model, minYear, zip);
+ return next();
 }
-// file where we actually execute the queries on the db
+
+// pgController.getCarGuruData = (req, res, next) => {
+//   const { make, model, minYear, zip } = req.body;
+//   res.locals.carData.carGuruData = carGuruScraper(make, model, minYear, zip)
+//   return next()
+//  }
+
+// APi (/asdasd/, Carsdotcomcontroller, cargurucontroller, => {
+//   res.locals.carguru, res.locals.carsdotcome 
+  
+// })
+
 
 // add controllers
 pgController.getSavedData = (req, res, next) => {
